feat: init dev plugin
