--- src/_index.ts
@@ -1,8 +1,0 @@
-function stylex() {}
-
-/**
- * @deprecated will be remove in next major version
- */
-const stylexPlugin = stylex
-
-export { stylex, stylexPlugin, stylex as edfault }
