+++ src/server/index.ts
@@ -1,0 +1,2 @@
+export { stylexDev } from './dev'
+export { stylexProd } from './prod'
